+++ CHANGELOG.md
@@ -3,6 +3,8 @@ CHANGELOG
 
 ## HEAD (Unreleased)
 
+## 1.0.0-beta.3 (2019-08-21)
+
 - When using StackReference to fetch output values from another stack, do not mark a value as secret if it was not
   secret in the stack you referenced. (fixes [#2744](https://github.com/pulumi/pulumi/issues/2744)).